feat(console): admin-level audit log + reusable component
- Move /api/[workspaceId]/audit-log -> /api/audit-log with workspaceId as
  an optional query param. With workspaceId: workspace-scoped, requires
  manageUsers in that workspace. Without: cross-workspace, requires admin
  (verifyAdmin). API also bulk-fetches workspace name+slug per row and
  returns it as 'workspace' so the admin view can render a workspace
  column.
- Extract the table into components/AuditLog/AuditLog.tsx. When workspaceId
  is undefined the component runs in admin mode: adds a Workspace column
  with name + link to /<slug-or-id>, and uses each row's per-workspace
  slug to build entity edit links.
- /admin/audit-log: new page that renders <AuditLog /> in admin mode.
- /[workspaceId]/settings/audit-log: now a thin wrapper around <AuditLog
  workspaceId={...} workspaceSlug={...} />.
- Add 'Admin Audit Log' entry next to 'Admin Workspaces' in the workspace
  selector menu (admin users only).
